feat: add user request limit config
