+++ docker-compose.yml
@@ -12,6 +12,7 @@ services:
       SESSION_TIMEOUT: 600 # 超时时间
       HTTP_PROXY: http://host.docker.internal:15777 # 配置代理，注意：host.docker.internal会解析到容器所在的宿主机IP，如果你的服务部署在宿主机，只需要更改端口即可
       DEFAULT_MODE: "单聊" # 聊天模式
+      MAX_REQUEST: 0 # 单人单日请求次数限制，默认为0，即不限制
     ports:
       - "8090:8090"
     extra_hosts:
